refactor: move pattern to utils
